fixing bug that was causing numba versions failing to execute from any directory other that dbscan
